CRRBV-22 - Fix for off by one bug near 32^2 elements The change to the clj version here is motivated by the corresponding code in Clojure's PersistentVector class, both for the persistent and transient versions, in its popTail methods.  I believe one way to think about why it is (cnt - 2) instead of (cnt - 1) here is that cnt is the number of elements in the factor before it is pop'd, so (cnt - 1) is the new number of elements after popping.  (cnt - 2) is the index of the new last element after popping.
The fix for the cljs version of the code is similar to the clj fix,
except that there is also a bug in tail-off that requires knowing
whether the vector is persistent or transient in order to return the
correct value.

An alternate fix, more similar to the Clojure version, would make
tailoff a method of a protocol that types Vector and Transient
implement differently.