CRRBV-13: Fix for the issue
The problem was with code that used a shift value for the wrong level
of the tree when updating nodes during assoc or assoc!  The fix is
specific to the clj implementation, and I believe the bug is, too.
Only the clj implementation has different cases for vectors of all
identical Java primitive types.