Fix several small things Although the Leiningen project.clj file is only a developer aid, and pom.xml is the official source for all things about the project, try to keep the version number in project.clj up to date with what is in pom.xml, in case a core.rrb-vector developer wants to run 'lein install'.
Fix a typo in a keyword of an ex-info error map.  Not a bug, but an
annoying inconsistency of names.

Reduce the run-time duration of the slowest cljs deftest, which is a
generative test.  This should help keep the typical pre-commit test
times on a developer's machine more reasonable, as well as reduce them
on build.clojure.org.  Developers are free to ramp up the durations of
these tests on their local development machines if they wish, of
course.