CRRBV-14: Add peephole optimizer for results of catvec
There is likely a cleaner improvement to the splice-rrbt function that
will give a result as good as this does, but this code basically just
does the normal splice-rrbt as part of the catvec implementation,
followed by a peephole optimizer that in some cases reduces the top 3
levels of the tree down to 2.

There is no known proof that this prevents the former problems of the
tree depth growing without bound when many catvec operations of
certain kinds are done, but in pretty extensive testing, no cases have
yet been found where the tree depth grows as it did before.