Large changes and additions to core.rrb-vector.debug namespace
The implementation has been split into a large nearly
clj/cljs-independent file, which could be merged into a single .cljc
file in the future, and a small clj/cljs-dependent file.

dbg-vec now takes an optional options map for controlling things like
maximum depth of the tree nodes to print, whether to show vector
element values, and a few others.  It also handles all kinds of
Clojure and core.rrb-vectors, including clojure.core/subvec return
values.

Several checking-* functions are defined, which in the past I have
used for somewhat quicker finding of the root cause of when problems
start.  They can be useful in test code, e.g. by calling
checking-catvec instead of catvec, because they check many conditions
on the return value before returning, that might not be immediately
obvious when examining the seq of the return value, for example.