Update test comments so they are current, and add extra-checks? aliases Add new deps.edn aliases, and jdo/sdo commands, to run tests with extra-checks? true from the command line, without having to edit any source code files first.  Also print the value of extra-checks? when the first deftest is executed.
Add doc string for debug/first-diff, and fix typo in another debug
namespace doc string.

Many comments in the test-common test namespace were written as I was
adding the failing tests, describing the failures in the present
tense.  Update them to say that they occurred with core.rrb-vector
version 0.0.14, since the failures no longer occur with the latest
version of the code.